requirements.txt
Appium-Python-Client
selenium
pyyaml
